Fix broken commit

+++ GrpcService/Services/Models/Calc/ExponentiationResult.cs
@@ -1,6 +1,6 @@
 namespace GrpcService {
     public partial class ExponentiationResult{
-        public ExponentiationResult(int resultValue) {
+        public ExponentiationResult(double resultValue) {
             Result = resultValue;
         }
     }
